bot/requirements.txt
# Runtime dependencies.
aiogram>=3.0,<4.0
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
pgvector>=0.3,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0

# Development and testing dependencies.
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
